[BE] refactor: 레디스 설정 파일 위치 변경  #763
